In preparation for space invaders ROM, made a directory specifically for it - separate from the CPU code, modified the main.c to reflect that it's strictly CLI, removed raylib, made rom_size to a uint64_t to properly check for sizes greater than 0xFFFF

+++ include/intel8080.h
@@ -19,7 +19,7 @@ typedef struct {
     memory_t    mem;
     registers_t regs;
     char        rom_name[1024];
-    uint16_t    rom_size;
+    uint64_t    rom_size;
     uint64_t    cycles;
 
     bool        is_halted;
